Bug#34506092 De-couple checkpointer thread from the log files
Description
-----------
The log.periodical_checkpoints_enabled is documented as enabling
periodical checkpoints (every 1s), but in reality
log_should_checkpoint() treats true as permitting it to also make
checkpoints whenever it would help free the oldest file, which creates a
tight dependency which isn't really needed, because file governor is
already rushing the oldest consumer in case it really needs to reclaim
oldest file, and checkpointer's consumer already implements a method
to create checkpoint in new file when requested to by file governor.

This isn't a bug, but it does force checkpointer to understand the
concept of "log file", take some low level mutexes, etc. just to
figure out if it should do a periodical checkpoint, which is not
needed or expected.

Fix:
----
Remove IB_mutex_guard files_lock call and other checks in
log_should_checkpoint to figure out periodic checkpoint
i.e. periodic checkpoints should be periodic, that is, dependent on
the time elapsed, not on some files occupation status.

Change-Id: Ie38f4cca60bf641c3369025ae2040bce4ea0eb1a