Bug#32860560 NDB_LOG_BIN: INCONSISTENT BEHAVIOUR WHEN SET DYNAMICALLY
Problem:
The --ndb-log-bin system variable, which is disabled by default in 8.0,
can be set dynamically but doing this doesn't have the desired effect.
The variable is only checked when the binlog thread starts and all
dynamic changes to it are silently ignored thereafter.

Ananlysis:
The --ndb-log-bin variables controls wheter or not the MySQL Server
should write NDB table changes to the binlog. When this functionality is
active the MySQL Server create subscriptions on all tables in NDB that
should be logged. It should not possible to turn this functionality on
or off at runtime, the MySQL Server need to be restarted. This is the
same as it's not possible to change the --log-bin variable at runtime.

Solutuion:
Mark the --ndb-log-bin variable as readonly, meaning it can not be
changed after MySQL Server has started.

Change-Id: I16895c8aea9b5574524ed68837182c217458badc