WL#16214 Export Query Results to Object Storage (Part 2/12)
AIM
=============
In the existing architecture all the data from HeatWave(HW) or
LakeHouse(LH) is displayed back to the user. The aim of the WL is to
export the result set to object store and just display the number of
affected rows to the user.

Architecture
==========================
The current syntax of SELECT statements needs to be modified to
capture the exporting to object store information.
The flow will be as follows:

User ==> MySQL client ==> MySQL Parser ==> MySQL query preparation
==> optimizer ==> QKRN ==> QCOMP ==> HW/LH execution ==> Arrow
library ==> object store.

When destination is object store, then instead of sending the data
from the compute nodes to plugin, the network chunks are converted into
the desired format using third party library Arrow. The converted data
is then dumped to the object store directly from the compute nodes.

Syntax
========================
The current syntax for copying data from MySQL:

<query expression>
INTO OUTFILE "<file name>"
[CHARACTER SET <charsetname>]
[ {FIELDS | COLUMNS} [TERMINATED BY 'string']
[[OPTIONALLY] ENCLOSED BY 'char'] [ESCAPED BY 'char'] ...]
[LINES [TERMINATED BY 'string'] [STARTING BY 'string'] ... ]

Here, <query expression> is an expression that qualifies a set of rows,
usually starting with SELECT.

We will be supporting both the syntaxes keeping OUTFILE to disk and
LakeHouse ENGINE_ATTRIBUTE in mind.

Syntax - 1 ( Non - json syntax)
==================
<query expression>
INTO OUTFILE
URL <uri>
[ FORMAT <format> ]
[ COMPRESSION 'string' ]
[ CHARACTER SET <charset name>]
[ HEADER <ON/OFF> ]
[ {FIELDS | COLUMNS}
  [ DATE FORMAT <date_format]
  [ TIME FORMAT <time_format> ]
  [ NULL AS <null_char> ]
  [ EMPTY VALUE <empty_value_char> ]
  [ TERMINATED BY 'string']
  [[[OPTIONALLY] ENCLOSED BY 'char'] / [ NOT ENCLOSED ] ]
  [ESCAPED BY 'char'] ...]
[LINES [TERMINATED BY 'string'] [STARTING BY 'string'] ... ]

Syntax - 2 ( Json syntax)
==================
<query expression>
INTO OUTFILE ENGINE_ATTRIBUTE =
'{"file": [{“par": ..., "prefix": ..}],
  "dialect":  { "format": "csv", "field_delimiter": ",",
                "record_delimiter": "\\n", "encoding": "utf8mb4",
                "escape_character" : "\\\\",  "date_format" : "auto",
                "time_format" : "auto","quotation_marks": "\'",
                "quote_all" :  false, "trim_spaces": false,
                "null_value" : "NULL", "empty_value" : "\'\'",
                "has_header": false, "compression": "snappy"}}'

(•) File uniquely identified by JSON entry. If the file already exists
then throw error.

(•) The json fields under key “file” cannot be empty. For example,
region/namespace/bucket/prefix cannot be left empty as they uniquely
identify a file.

(•) There will be default values for the fields under “dialect” section.
For example, if no field_delimiter is specified, the default format
will be “ “.

(•) This syntax is similar to LH ENGINE_ATTRIBUTE and hence easier
to copy paste during load

Dealing with Default values :
======================================
The same default values have different names for LH load and Innodb
OUTFILE. For example,
• field_delimiter is same as FIELDS TERMINATED BY
• escape_character is same as FIELDS  ESCAPED BY
• record_delimiter is same as LINES   TERMINATED BY
• quotation_marks  is same as FIELDS ENCLOSED BY
• encoding is same as CHARACTER SET

Whenever HW/LH defaults vary from MySQL (e.g., delimiters etc.),
we will conform to LH defaults here, as in this version we will
be supporting export only from LH/HW.

Computing the number of affected rows
=========================================
In case the results are exported to object store, the actual result
set is exported to the object store but the total number of output
rows have to be displayed in the client.

In addition to sending the data to the object store, each server
will transmit the number of rows in the output network chunk back to
the plugin. The plugin will gather the number of rows information
from the chunk metadata and calculate the total number of rows.
Since the actual data no longer needs to be displayed, the vector
buffers are removed from the chunk before sending it back to RDBMS,
and the payload size is adjusted accordingly.

In the existing Query_result_send, the MySQL client itself computes
the number of rows based on the result set. However, in this scenario,
the plugin only receives the chunk metadata. Therefore, a new class,
Query_result_to_object_store, is introduced. Once all the chunks,
followed by the end-of-query message, are received by the plugin,
it will simply send the total number of rows to the MySQL client.

Execution in PRIMARY and SECONDARY engine
=========================================
Exporting results to object store can be executed only in Heatwave
and LakeHouse. In case the query cannot be executed in HW/LH , then
it cannot be executed in Innodb either since Innodb does not support
this functionality yet. The error message will be "ERROR HY000:
'OUTFILE to object store' is not supported

Supported Output formats
======================
In this WL, we will be supporting exporting of data only in
CSV (default), TEXT and parquet.
TEXT format will be supported for exporting to both disk and
object store destinations.

Compression Algorithm
===========================
The compression algorithms supopported so far are
 "uncompressed", "snappy", "gzip", "brotli", "zstd"
 "lz4", "lz4_frame", "lzo", "bz2", "lz4_hadoop"

New Error code
=========================
A new error code ER_INCORRECT_PARAM_ENGINE_ATTR
is introduced.
This error is thrown as a warning when skip_rows,
is_strict_mode, allow_missing and constraint_checks
are part of the ENGINE_ATTRIBUTE JSON string of OUTFILE.
On the other hand, null_value, empty_value and
compression are ignored in the ENGINE ATTRIBUTE json string
during LH load.

Change-Id: I4ca64b9383f96722e5912ad7898476d87b92190f