WL#16701: Collect simple statistics in row iterators
Refactor the RowIterator class hierarchy so that some simple
statistics are collected during execution. The iterators now collect
the following statistics:

- The number of times the iterator is initialized/re-initialized.

- The number of rows returned by the iterator.

- The number of times the iterator was read all the way to EOF.

This is done as a preparation for other changes that aim to improve
the selectivity estimates in subsequent executions of the query based
on the actual selectivity observed in earlier executions.

Change-Id: I1f2bb75dec6e205c77cdb0e617aaed3f26758479