Bug#36968010 Router accepts the config file when the mysql_rest_service.router_id is missing
In the current implementation when the MRS is bootstrapped, there is an entry
in the mysql_rest_service.router created and the router_id is stored in
the configuration file in [mysql_rest_service].router_id.

However if the Router is launched with the config file that is missing
the router_id it would just ignore it and proceed with not reading the
options and not doing the periodic updates to last_check_in field.

This was somewhat acceptable before WL#16483 Add support for MRS service
'in_development' feature. Now the router_id is used to determine if a
given service instance should be exposed by the running router instance
or not. Hence this field should be required.

This patch makes the router to fail to launch if
[mysql_rest_service].router_id is missing in the configuration.

Change-Id: I81f818d8582fe832d6ae54e1fe2843dec55de9ea