BUG#36987698 Router does not honor the url_host configured for service in the metadata
When the MRS service is configured, the user can define the host this service
should handle request on. This should then be matched with the host in the
request headers when handling a request.
Currently the MRS ignores that value. As a result if there is a request
incoming, only the service name is checked if it's matching the existing
service. If it does the request is handled by a given service.
One can configure multiple services with the same name but for different
url_host entries. In this case the current MRS behavior is random, the first
service that matches the name is handling the request, the host header is ignored.

This patch changes the behaviour so that:
- The host from the headers of the request is compared to the service url_host
  configured. If those to match exactly the given service is used to handle the
  request.
- If there is no service with a given name and matching url_host configured
  the one with empty url_host configure is used to handle a request
- If there is no service with a given name and matching url_host configured
  end no service with a given name and empty url_host configured, the request is
  not handled (404)

Change-Id: If38255710cc150193299bdb06bd51f9a8f631957