WL#16214 Export Query Results to Object Storage (Part 1/12)
New Error code
=========================
ER_CLAUSE_IGNORED : When an input argument in OUTFILE is
ignored as it is not yet supported by the backend.

ER_INVALID_CLAUSE_IN_CONTEXT : When an input argument
in OUTFILE is invalid in while exporting query result
to object store or disk.

ER_INVALID_CLAUSE_COMBINATION:  When an input argument
in OUTFILE is invalid in combination with the rest of the
inputs.

Change-Id: I4b54b8754106edc291931be333a3b8e53be88f8a