Bug#36998710 MRS does not stop the REST service when the standalone node becomes read-only
In the current implementation the MRS does not check if the
node that is has configured for RW operations is in fact not read-only.
As a result the service is not stopped in that case and the
metadata update periodic INSERT statements are kept executing on that node.

This patch changes the behavior so that the MRS stops the REST service
once it discovers that the destination server that is configured as RW
became RO. This is done only if the destination is standalone server(s).
The behavior for the Cluster monitored with metadata-cache does not change.

When the node becomes RO the MRS keeps checking if it is back RW, and
only then restores the service.

Change-Id: I1079966173b9f3cb69bd50c7023d55bbfba1ef44