Bug#36995669 - Out parameters format in procedure response is incompatible with ORDS
Description
===========
The response generated for db-object-procedure, should be similar to ORDS response.
The key difference between MySQL and Oracle is that MySQL allows returning resultsets
from the procedures. Thus output parameters and resultset should be contained
in root document in two differen fields. The JSON object should follow this layout:

  {
    "resultSets": [
      {
        "type": "ResultsetName1",
        "items": [{"resultset1-row1-field1": "value1", "resultset1-row1-field2": "value2" },
                  {"resultset1-row2-field1": "value3", "resultset1-row2-field2": "value4"]
      }, {
        "type": "ResultsetName2",
        "items": [{"resultset2-row1-field1": "value5", "resultset2-row1-field2"...}...]
      }],
    "outParameters": {
      "output1": "foo",
      "output2": "foo"
    }
  }

Fix
===
Aligned the REST response generated from procedures to the new format.
The new format is will keep "metadata_" field, in JSON object describing
a resultset. It was removed for 'outParameters'.

Change-Id: I5c986bc06a117059d2ad5d93fcdf91d04808c410