WL#16515 Allow configuring [routing] endpoints to not accept the external connections
This patch allows the [routing] section to be configured so that
there is no acceptor for the client connections neither on TCP port
nor the Unix socket.

[routing]
accept_external_connections=0

This is useful for MRS use-case where the routing is used only to
feed the destinations to the mysql_rest_service plugin, and we do
not want to expose the accepting ports.

Change-Id: Iae28144f6d8b5fd2c755cebb96711e140b68b548