Bug#37100804 avoid strlen() in rapidjson calls
Passing std::string.c_str() to a rapidjson::Value() leads
to a call of strlen() to get the length of the string.

rapidjson supports taking a std::string as Value directly if
RAPIDJSON_HAS_STDSTRING is defined.

Passing the std::string to Value directly, instead of its .c_str(),
avoids the strlen() as the std::string's .size() can be used directly.

Change
======

- enable std::string support in rapidjson
- replace all .c_str() inside Value() with passing the std::string directly
- call new_doc.get_allocator() once.

Change-Id: I036846c0de3abe2580490e2ebbdd883fc46786a3