BUG#37463237 Make innodb.sdi agnostic to ndb_binlog_index
Problem:
The test innodb.sdi runs only when MySQL has been compiled
without support for NDB. Since the default is to compile MySQL with
support for NDB, the test does not run so often and changes it intend to
test may go unnoticed.

Analysis:
The difference between MySQL compiled with suppport for NDB
is that the mysql.ndb_binlog_index table are created when the system
database is bootstrapped. This means that any test which show or
otherwise work with the tables in the MySQL system database need to
adapt itself to this tables existence. There are different strategies
used by tests for handling this, one of the most common is to simply
mask out the table and or it's columns from ending up in result files.

For sdi.test the easiest is to always make sure to test with a
bootstrapped system database that does not contain the
mysql.ndb_binlog_index table. This is done by simply dropping the table
from the bootstrapped system database that the test creates. The only
side effect of this change is that the table_id counter has already been
incremented by one and thus subsequent tables created get a higher than
expected table_id. However for this and similar reason the sdi test
already masks out all table_id's in the sdi output - except in one place
where the regex to mask table_id= for a table that has another
se_private_data property before the table_id= property.

Solution:
Remove ndb_binlog_index from system database bootstrapped by test.
Fix the masking of table_id= also when there are prior properties, this
is done with a new regex after the others not to affect all saved .json
files.
Remove usage and existence of the have_only_innodb.inc include file that
checked for this condition.

Change-Id: Ic3c43c1f084a108f497cae882432ac03f2485661