Bug#37100619 Router fails to use mrs schema if minor version is greater than the one it knows
When you bump the metadata version from 3.0.0 to 3.1.0 and try to run
the Router that was bootstrapped for MRS it will refuse to read the metadata
with the following error:

ERROR [0x16bcd3000] Can't refresh MRDS layout, because of the following error:
Unsupported MRS version detected: 3.1.0.

This should not be the case. The minor version bump should be non-breaking and
backward compatible, so the Router should use the metadata if the minor version
is higher than the one it knows.

This patch changes the logic so that MRS does not complain if the minor
version is higher than the one it konws (has embedded).

Change-Id: Ic793dcb6a29717a4c19568f9d9e00a95a5297b07