MRS endpoint refactoring
Description
===========

Remove `object` and `object_schema`, instead create endpoint class, which represents each part
of the URL:

     HTTP://HOST/PATH1/PATH2/PATH3
            ^     ^      ^    ^
            |     |      |    db_object_endpoint
            |     |      |
            |     |      db_schema_endpoint
            |     |
            |     db_service_endpoint
            |
            db_host_endpoint

Each endpoint is reponsible for allocating specific it the path, URL handler, which
are reposnsible for generating HTTP reponses for given request.

For example, the db_object_endpoint is going to create following handler:
  * option-file handler (options column, is allows to specify in json additional
    files that may be anchored to db-object-path).
  * objects-metadata catalog
  * object REST handler

Other endpoints share similar handlers to the user.

Change-Id: I46644422b727ea646dc3061e7fa8e855b78b9206