BUG#35206392: Adding secondary Node, make existing secondary nodes to start lagging
Background
----------

There is a Ticker Manager, which ensures that transactions are
committed on the correct side of a view_change. I.e., transactions
certified before the view_change commit before the view_change, and
transactions certified after the view_change commit after the
view_change. This is achieved as follows:
 * The view_change waits for preceding transactions (according to
   global order) to begin committing, before the view_change begins
   to commit.
 * Transactions following the view_change (according to global order)
   wait for the view_change to begin committing, before those
   transactions commit.

On an inbound replication channel, there is a Commit Order Manager,
which ensures that transactions are committed in the same order they
are received. This is achieved by making each transaction wait for
the preceding transaction (according to receive order) to begin
committing, before it commits.

Problem
-------

Whenever these two Managers require different orders, it results in
a deadlock. For example, suppose the following happens:
 * The inbound channel receives T1 before T2.
 * T2 is certified first. Then a view_change occurs. Then T1 is
   certified.
This leads to a deadlock:
 * When T1 is about to commit, it will invoke the Ticket Manager,
   which waits for the view_change.
 * When the view_change is about to commit, it will invoke the Ticket
   Manager, which waits for T2 to commit.
 * When T2 is about to commit, it will invoke the Commit Order
  Manager, which will wait for T1 to commit.
Therefore, this results in a wait cycle, i.e., deadlock.

Solution
--------

We detect the deadlock, enforce the Ticket Manager order and ignore
the Commit Order Manager order. This has the following user-visible
consequences:
 * It may violate replica_preserve_commit_order near
   View_change_log_events. In other words,
   replica_preserve_commit_order no longer provides any strict
   guarantee on an inbound channel on a GR primary.
 * replica_preserve_commit_order still ensures a that transactions
   are likely ordered correctly, with exceptions only around
   View_change_log_events.
 * This suffices to avoid GTID gaps almost always, which ensures that
   maintaining gtid_executed is fast.
 * But the user cannot rely on transaction order to be always
   preserved for inbound channels on GR primaries.

Change-Id: Ibbbae77dadc9db79381decc535713d2962d7ffa9