Bug#37135458 - Unable to call REST function/procedure with "optional" parameters
Description
===========
As an MRS user, I expect to be able to call a REST function or procedure by
submitting a PUT request where the body is a JSON document where the keys
correspond to function parameter names or procedure "IN" or "INOUT" parameter
names and their corresponding values.

Additionally, I expect to either be able to omit certain parameters
that I deem to be "optional",

Fix
===
Do not throw exception, when MRS detects that parameter was not
set by the user.

Change-Id: Ibc26b52f67c8c2f0747075b04e8c5971111ca8e4