WL#16409: HeatWave Offload Debuggability (Phase 5).
Bug#36531144: MySQL Item - No m_pos parser info available in some MySQL items that is useful in backtracking to original query.

MySQL code changes related to WL#16409.

Push POS information to function items to allow gathering of
information that is useful for pin-pointing query fragments
in non-offloading HW queries (for debuggabilty phase 5
improvements).

Parsed text context was missing for comparison functions, such as
equality, greater than and less than.

This has been fixed by using POS-supplied constructors for items
representing these functions.

Since the POS-enabled constructors do not implicitly add the items
to the current list of items, this had to be done explicitly.

We also had to add an explicit call to a new function
Item::set_contextualized() in order to ensure that all item objects
are properly contextualized, which is monitored with assert statements.

Propagation of item properties had to be added explicitly in
Item_bool_func2 constructor, since this is not done implicitly for
POS-enabled constructors.

Change-Id: I7699fe844523f9e950a9586d7664c48be2462b00