Fix for random MTR tests failures where the test timeouts waiting for the MRS to report noticing added objects.
This patch is a bit hacky way of making sure that MRS fully reinitializes the
state from the metadata between 2 different tests run on the same MTR worker.

We do it by forcing some high audit_log.id that would cause the MRS to
notice inconsistency on the next test run.

Change-Id: Iffe84b1499d8474a1454ff88b6196c73b5ca8430