Bug#37174836 SDK fails on CREATE REST SERVICE localhost/sdkService
When the MRS service is configured, the user can define the host this service
should handle request on. This should then be matched with the host in the
request headers when handling a request.

In the current implementation the host must match exactly.
So if the service was created for host HOST1 and the request has
HOST1:8848 as hostname it will not match.

This patch changes this behavior so that if there is no exact match
the MRS will check if the url_host from the request is not in the
HOSTNAME:PORT format. If that is the case, it will check for the handler
registered from HOSTNAME host and use it if found.

Change-Id: I8814158b28d95e3e1427b78e1a2a847210104247