BUG#37008345: Stop replica blocks after reporting the replica coordinator and worker threads
* Background

After the fix for Bug#36507020 the receiver thread will wait outside
transaction boundaries whenever there is no space to queue according
to the relay_log_space_limit variable.

On the receiver thread start though, there is no way of knowing when
we are inside of a transaction or not, and source events can be logged
after an unfinished transaction in the relay log.

* Problem

Whenever the receiver thread stops after queueing half of a
transaction, it can be that the relay logs that are present on a new
start have a total size above relay_log_space_limit. This prevents the
receiver thread to queue Rotate or Format description events that came
from the source.

The FD event is important cause it is this event that signals the
applier that the half transaction it is scheduling is incomplete in
the relay log and should be rolled back.

The applier never receives this event, as the receiver thread is stuck
waiting for space and remains waiting. The applier never reaches the
point, after scheduling a full transaction or rolling back a
incomplete one (as the relay log is at the end), that would cause a
relay purge and unblock the receiver thread.

In summary the applier thread is executing half a transaction and
never reads the event that signals that rest of the transaction will
never be delivered, and for that reason will never unblock the
receiver thread that is trying to send this signal but can't cause it
lacks space (due to relay_log_space_limit)

There was also an issue with the applier reader code concerning
transaction boundaries. Transaction boundaries are tracked using
different means and a different API in single-threaded and
multi-threaded mode and this code used the API for the single-threaded
mode, even in multi-threaded mode.
This issue eventually lead to the STOP REPLICA command not being
responsive as the SQL thread could not be killed.

* Fix

Make the receiver thread ignore the relay_log_space_limit for Rotate
and Format Description events that come from the source. This way the
applier will eventually progress and clean up the old relay logs that
are taking space.

We use the correct API (depending on whether the applier is in
single-threaded or multi-threaded mode) to check if the applier is on
a transaction boundary or not.

Change-Id: Ia26289c144b3b194980536d6aa03b91dae1d799a