Bug#36890891: If SQL in constraint fails, table becomes unopenable / unfixable
When SQL functions are improved, they may throw SQL errors in situations
where they previously did not. If this happens in a table's constraints,
default expressions, partitioning, or virtual columns, the table can not
be opened. This will prevent both analysing the problem (e.g. with SHOW
CREATE TABLE) and addressing it (e.g. with ALTER TABLE ... DROP ...).

These issues can be addressed in two ways:

a) On server upgrade, we scan the data dictionary for tables that
   use any of the aforementioned features. We will then try to
   open such tables, and if we fail to do so, alert the user.
   This patch addresses this.

   --check-table-functions=WARN|ABORT

   WARN  logs a warning for each table we couldn't open.
         For a sufficiently high --log-error-verbosity, it
         also logs a note with a streamlined table definition
         listing only the expressions that potentially contain
         SQL functions.

   ABORT logs the same warnings as WARN, but aborts the server
         upgrade if any issues were found. This is the default.

   The default behaviour is to abort so the user can fix the
   issue using their "old" server before upgrading to the new one.
   The non-aborting variants enables variant b) below, namely
   flagging the issues, but letting the user continue to
   interactive mode so they can fix the problem

b) SHOW CREATE TABLE and ALTER TABLE are more forgiving of certain
   SQL errors during table opening to allow for analysing and
   fixing of these issues.
   Bug#37009318 addresses this.

Change-Id: I5e4687dc7b3435a3c59e41edcd203de89bd7e060