Fix for authentication_endpoints.test failing on slow machines
The test checks the authentication apps returned by the MRS via REST
request. The problem is that if there is more than one returned their
orded in JSON array depends on the order the server ruturns the rows in the
query which has no ORDER BY clause. This is somehow random.

To fix it the schema is used to validate the result regardless of the
order instead of direct matching via result file.

Change-Id: Ide646b07c57268c5ab2a19dc5fb7c7e5a50120aa