Bug#37194991: joining 8.4 member into an 8.0 group get stuck in recovering state
* Background

Each transaction has two logical timestamps, stored in the GTID event:
sequence_number and last_committed. Before the coordinator schedules
the current transaction, it waits until the transaction whose
sequence_number equals the current transaction's last_committed has
committed. We call this "wait on the dependency". Waiting on a
dependency ensures that conflicting transactions don't execute in
parallel.

The scheme works as long as sequence_number increases by 1 from one
transaction to the next, and each transaction has last_committed <
sequence_number. So, there are three ways that the numbers can be
inconsistent:

I1, last_committed >= sequence_number.
I2. The sequence number is less than or equal to the sequence number
of the preceding transaction.
I3. The sequence_number increases by more than 1 from one transaction
to the next. We call this a "gap in the sequence".

If I1 or I3 occurs, there may be no transaction with sequence_number
equal to the value of last_committed for the current
transaction. Therefore, if the coordinator would wait on the
dependency, it would wait forever. To prevent such deadlocks, the
coordinator checks for all three cases. If it detects I1 or I2, it
reports an error and stops. If it detects I3, it waits for all
previously scheduled transactions to commit (rather than wait on the
dependency). We call such a wait a "barrier".

The source server resets sequence_number to 1 when it rotates the
binlog, so then it expected that case I2 occurs.
Therefore, if the coordinator detects this situation, it
- suppresses the 3 checks,
- and forces a barrier.

It uses two different conditions to determine that the source has
rotated the binlog or we want force a barrier:

C1. there is a format_description_log_event originating from the source;
C2. there is a GTID event having special values for last_committed and sequence_number.

It suppresses the 3 checks (I1,I2,I3) if (C1) holds.
It forces a barrier if (C1) || (C2) holds (or rather, this was the intention: see below).

* Problem

The condition to force a barrier was effectively just (C2), rather
than (C1) || (C2) as intended. Therefore, if there was a
format_description_log_event, and a gap, but the GTID event did not
have special values for last_committed/sequence_number, then the
coordinator would wait on the dependency, and because of the gap,
could enter the deadlock situation described above.

This condition could occur in relay logs produced by GR in server
versions 8.0.37 or older, in which BUG#36280130 had not yet been
fixed. BUG#36061197 also contributes to the issue as it made C1 more
common for GR scenarios.

* Fix

We change the logic to:

If C1 does not hold, report error and stop if the logical timestamps are wrong in the sense of I1 and I2.
Unconditionally, force a barrier if the timestamps are wrong in the sense of I3, or if C2 holds.

Change-Id: I6fd990e60f00e3f680dbbb619c99484df7c4ac63