Bug#35006212  Unnecessary conversions performed for end range check
Symptom:
Occasionally conversion of data from InnoDB to MySQL format breaks
assumptions made by code about the state of row data. A known example is
a scenatio where a LOB field subject to conversion is in a "being
modified" state in a transaction isolation level other than READ
UNCOMMITTED. This fails a debug assertion - in this particular case
there is no undesired behavior in release builds.

Cause:
The scenario occurs when a record is converted to MySQL format prior to
performing an "end range check" optimization as part of an index range
scan. The record used for the check is the newest version of the record,
potentially uncommitted and not in the view of the current transaction.
This is generally reasonable as the end range check uses only key fields
of an index, but InnoDB converts all fields leading to unexpected
scenarios and states in the conversion functions. In the LOB scenario
described above the record in a transitional state may be seen when
locks are released for the purpose of writing a redo log when the LOB
being written is sufficiently large.

Fix:
Perform the end range check using InnoDB's tuple comparison.
Range-based tuple comparison was disabled for a number of cases,
including partitioned tables and skip scan scenarios, using
the is_reading_range flag. This flag is no longer needed and is
being removed. The result is that some range-based checking and locking
not directly related to end range checks is also applied more widely,
in particular to partitioned tables; in some scenarios this will reduce
row-level lock congestion.

Change-Id: I3d4649d27797812aed29fdeec2461bc4d4b97edc