Bug#37200056 - MRS service authentication-logout endpoint returns http status different                than status presented in json.
Description
===========
When sending a logout request to service authentication endpoint, the response is same
for authenticated user, and not authenticated. The response is following:

    GET:    /myService/authentication/logout
    STATUS: Unauthorized
    RESPONSE BODY:
    {
        "message": "OK",
        "status": 200
    }

It may be confusing to the user because the HTTP status is different from the status
in json-response. The main one (http) says that response is faulty, and json says ok.

Fix
===
The HTTP status should be the same as with the json response.

Change-Id: Ie8147456d6e9f92faf2850c1e1a54ae5f33d5d46