Bug#37372130 Add version number to JSONv2 EXPLAIN
Add new field "json_schema_version" to the top level of JSONv2 EXPLAIN
output. Initial value is set to "2.0".
This value is there to make it easier to know if there has been any
changes to the format of the JSONv2 EXPLAIN output that could warrant
code changes from the user.

Change-Id: Ifeae31952a6533fa6b955ffdcf28f3d14a992c8e