Bug#34361287: Select with CAST and having differs between OO and HG Bug#34551116: Query with 100 rows from OO, 0 rows from HG
Some queries that performed arithmetic that returned a DECIMAL value
returned different results depending on which query plan was chosen.

The problem was caused by the arithmetic operation returning a DECIMAL
whose internal representation had a higher precision than the metadata
of the expression indicated. If the value went through a
materialization step, the extra precision would be removed when
storing the value in the temporary table, and the result was as
expected. If it was not materialized, the extra precision was kept,
and different results were seen.

Fix by rounding the DECIMAL value to the resolved precision before
returning it in Item_func_numhybrid::val_decimal() and
Item_func_numhybrid::val_real(). (The rounding is already done in
Item_func_numhybrid::val_str().)

Change-Id: Ic514b3909d06914cd6e60646ad3c41b25256be5f