WL#13052: Transform table subqueries using derived tables
    Stage 2: Implement the new transform.

A new class Item_bool_if that is a simplified version of class
Item_func_if is implemented. It can only return boolean values
(TRUE, FALSE and UNKNOWN) and is used to make results that correctly
return UNKNOWN values.

The value_transform associated with class Item_exists_subselect has
been simplified to only BOOL_IS_TRUE and BOOL_IS_FALSE, since EXISTS
and NOT EXISTS never return an UNKNOWN result. This made it possible
to simplify the code slightly. It also had the pleasant sideeffect
that EXPLAIN for NOT EXISTS now prints (not exists ...) instead of
(exists (...) is false.

Item_exists_subselect::is_derived_candidate() is extended so
that it exactly determines the cases where the new transform is
possible. See worklog description for details.

Query_block::transform_table_subquery_to_join_with_derived()
contains the actual implementation work. The new transform relies
upon the basic functionality already implemented in this function,
with support for decorrelation of inner WHERE clause, transformation
of subquery into derived table, synthesis of new join condition.

The function is extended with support for quantified comparison
predicates, according to conditions specified in the worklog.
The subquery is converted to an aggregated subquery (if necessary),
and the required aggregate functions are added as selected fields
of the derived table. Based on actual operation and properties of
involved expressions, MIN or MAX of the involved selected expression,
a count of the number of rows in the subquery, a count of NULL values
for the selected expression, as well as a count of number of distinct
values are selected.

A new condition is then generated to combine information from the
fields selected from the derived table, plus the left-hand expression
of the quantified comparison predicate. Finally, the generated condition
is installed as a join condition in the join between the outer tables
and the derived table (if the predicate is placed in the WHERE clause),
or as selected expression (if the predicate is placed in the select
list).

The function has also been enhanced to support transformation of
correlated subqueries, with outer references in equality predicates in
the subquery's WHERE clause. We tranform the subquery into a grouped
subquery for this purpose, where the subquery is grouped on the outer
reference expressions.

Includes fix of bug#37529060.
Includes fix of bug#37533056.

Explicit testing for the hypergraph optimizer has also been added.

Change-Id: Ie82c0a36c1b12e047bb7a5e32bdff224558684f4