Bug#37349412 - Unexpected MRS auth error if there are multiple auth apps registered for a service
Description
===========
As an MRS user, I expect that if I try to authenticate in a REST service
using an MRS authentication app, provided that the corresponding REST user
exists and the password is correct, I should get a successful
response with the corresponding access token.

Although this seems to be the case when there is a single MRS authentication app
registered for a given service, that behavior falls apart once I register an
additional authentication app of any type (MRS, MySQL Internal or OAuth).

In this case, instead of starting the underlying SCRAM exchange,
the MySQL Router sends back a "307 Temporary Redirect" response that redirects
the client to a forbidden resource.

Fix
===
The MySQL Router should allow and abide by the SCRAM exchange in any circumstance,
as long as it is using an MRS authentication app. In this case, it should eventually
return back a successful response containing the access token regardless of the number
of auth apps that are registered for the service.

Change-Id: I2eb7da57bdfa082771f717ea973e32c26df520f2