Bug#37292905 Can't launch MySQL Router in MTR tests.
MySQL Router source code is part of the mysql-server repository
for many years at this point.
As such it would be convenient to be able to use MTR launcher
to also test the Router binary or/and the Router server
(or even cluster) integration. MTR has many useful features
like request/response flow where the output is matched
against the pre-recorded result files. This is specifically
useful in context of a new MySQL Router project which is
"MySQL REST service".

This patch integrates the MySQL Router with MTR and introduces
the Router suite containing basic example Router test
with InnoDB Cluster.

Change-Id: Icddff4fb80008dc020f0f56d86f574d48d262949