Bug#35989799: Discontinuity in row estimate for ROLLUP in hypergraph
Post-post-push fix.

In the post-push fix, EstimateAggregateCost() was changed to set the
row estimate of an aggregation access path to unknown if the input row
estimate was unknown.

This also affected aggregation access paths for implicit grouping,
where we know exactly one row will be returned regardless of the
number of input rows. If the number of input rows was unknown, the row
estimate for the aggregation was set to unknown after the post-push
fix. Such access paths should still have a row estimate of one.

main.subquery_bugs failed because of this.

Fixed by moving the check for unknown input rows after the check for
implicit grouping.

Change-Id: Ie8eda67eb7bbbe7b27ddd793a7c4a626fa961432