Bug#37273221: Hypergraph: Inefficient join order when using LIMIT
For inner hash joins, the hypergraph optimizer currently only
considers the join order which gives the smallest number of rows in
the hash table. This always gives the lowest estimated total cost for
the hash join with the current rudimentary cost model for hash joins.

If the query has a LIMIT clause, however, the whole result of the hash
join doesn't necessarily have to be read, so it might be beneficial to
pick a join order with a lower init cost in order to reach the limit
sooner, even though the cost of performing the entire join is higher.

This patch makes the hypergraph optimizer explore both join orders for
inner hash joins that are under a LIMIT, so that the one that results
in the lowest estimated total cost for the query can be chosen.

Change-Id: I744e3188f493fd2c211e5d10934c9222905fe9bc