Increase max. buffer size for SP OUT params to 1MB
Bind buffers for text/blob OUT params called via prep stmt API
was limited to 1000, which is insufficient for GenAI use cases
that pass around large amounts of text.

Limit was increased to 1MB.

Possibility of starting with a smaller buffer but retrying with
larger ones if not enough should be evaluated in the future.

Change-Id: I3221368632598805eb3709295754638ed630a22f