Fixed error calling SPs with JSON that has a {} in it
Calling a SP passing a JSON structure with a {} in it
(e.g: {"prompt": "hello", "options": {}}) was failing and
returning 500.

Issue was that when converting input params from JSON to SQL,
only literals and arrays of ints were being handled. Code was
changed to just always pass arrays and objects as a string to SQL.

Change-Id: Id3ea5f1f723ce5aa82da81c447d95524c36d2133