Bug#37381478 User for MRS bootstrap needs priviledges for mysql_innodb_cluster_metadata even if no cluster
The minimal set of priviledges for the mysql user performing MRS bootstrap
on standalone serever should be as follows:

GRANT SELECT, INSERT, UPDATE on mysql_rest_service_metadata.* to mrs_bootstrap_user;
GRANT mysql_rest_service_data_provider TO mrs_bootstrap_user WITH ADMIN OPTION;
GRANT CREATE USER ON *.* TO mrs_bootstrap_user WITH GRANT OPTION;
GRANT mysql_rest_service_meta_provider TO mrs_bootstrap_user WITH ADMIN OPTION;
FLUSH PRIVILEGES;

However if such user is used it fails with:

Error: Error executing MySQL query
"SELECT * FROM mysql_innodb_cluster_metadata.schema_version":
SELECT command denied to user 'a'@'localhost' for table 'schema_version' (1142)

The problem is that without an access to mysql_innodb_cluster_metadata the
Router can't tell the difference between this schema not existing and the user
lacking the rights for it.

This patch changes the behaviour so that now the bootstrap continues as if the
mysql_innodb_cluster_metadata did not exist and there is a warning printed on
the console suggesting re-bootstrap with an user that has rights to that schema
if that is not a desired behaviour.

Change-Id: I52110e626c385bcb9154416f6a0a3c84b61056e5