Bug#37373959: Add rough InnoDB cost into temp-table-aggregation [postfix WL#16117]
If the estimated temp table size crosses @@tmp_table_size, bump up the
in-memory temp table cost by 5 times, for temp-table-aggregation path.

WL#16117 calibrated the cost of temptable aggregation using the
TempTable engine, which is used by default for internal temporary
tables when their size is below the threshold @@tmp_table_size (16 MiB
by default). Temporary tables that exceed this size limit instead use
the InnoDB engine. Experiments have shown that using the InnoDB engine
for temptable aggregation is approximately 5 times slower compared to
using the TempTable engine.

Using the cost estimates for the TempTable engine in cases where the
InnoDB engine was actually used to perform temptable aggregation was
the cause of some regressions in TPC-H benchmarks. This was due to the
optimizer underestimating the cost of temptable aggregation, and
picking temptable aggregation when streaming aggregation would have
been faster.

This is a quick fix to account for the cost of temptable aggregation
using InnoDB tables in the case where they fit in the buffer pool.  A
future worklog needs to properly calibrate the InnoDB cost overhead,
for any materialization path in general.

Change-Id: I2752782bc5175d1bce87f301c6dcdc6b4f820ada