Bug#37392546 MRS does not stop the REST service when the standalone node becomes offline mode
Patch for "Bug#36998710 MRS does not stop the REST service when the standalone node becomes read-only"
changed MRS behavior so that it stops the service when the standalone destination node becomes read-only mode.

The same behavior is expected when the node becomes offline.

Currently when this happens the MRS keeps logging every ttl:

Can't refresh MRDS layout, because of the following error:
Error connecting to MySQL server at /var/run/mysqld/mysql.sock: The server is currently in offline mode (3032)

This patch changes the behaviour, now when the MRS monitoring thread notices the offline mode, it
will disable the service and keep checking for the offline mode to be off again.
The error messge about the offline mode gets logged only once.

Change-Id: I7b80bd7f52dd640516456635648b3f1e3f496c28