Bug#37382579: Hypergraph crash with JSON_TABLE and escaped characters
When using the hypergraph optimizer, sorting the result of joining a
table and a table function could give wrong results or other erratic
behaviour if the sorting used row IDs. It would typically use row IDs
because of BLOBs in the result.

The problem was that the row IDs that pointed into the temporary table
with the results from the table function, are invalid if the table
function is scanned again after the row ID was fetched. They are
invalid because the temporary table is always emptied and repopulated
when the table function is rescanned.

The solution is to materialize the result of the join in another
temporary table before the sorting, and let the sorting store row IDs
that point into that temporary table instead. This is achieved by
marking the table function access path as unsafe to get row IDs from
if scanned more than once.

Change-Id: I77572aa708739aadd8bd220ea4f8ebc73df10fdc