Bug#35694966 Move rapid.ml_* tests into a separate test suite
Problem
=======
The ML MTR tests are timing out on PB2 and Jenkins pipelines.
They have different resource requirements than the rapid tests.
We can clearly see that the tests are under utilzing the compute and
memory capacity.

Solution
========
We make the following changes to speed it test runtime for ML tests.
We improve the total runtime from 6 hours to 1 hour.

- We increase the number of workers for the ML driver to 4
- We seperate the ML tests into their own seperate test suites namely
rapid/ml and rapid/genai
- We divide the big tests into smaller tests to improve parallelism
- We reduce the GenAI tests runtime by reducing the expected input and
output for the LLMs

Change-Id: Ifc07e23945d90547c4bba172e055fcfb07ec3d0f