BUG#37398919 - Errors are not correctly handled in udf_handler::add for aggregates
Problem:
--------
In udf_handler::add, when a UDF is an aggregate, errors returned by the
UDF are being treated as NULL, allowing the UDF to continue processing
for all rows. This behavior may not be ideal, as it delays the detection
of issues and allows unnecessary processing.

Solution:
--------
Modify the behavior to immediately return the error to the caller when
a UDF error occurs. This ensures that the query terminates promptly,
reducing wasted computation and providing quicker feedback to the user.

Change-Id: Ia81df67304ddbcde4c90bf88f5401f3cb375d7ab