Bug#36630355: Assertion `false && "Inconsistent row counts for different AccessPath objects."' failed.
Some queries using STRAIGHT_JOIN failed with assertions caused by
inconsistent row count estimates when running with the hypergraph
optimizer.

The inconsistency was seen because one predicate (the t1.d = t3.d
predicate in the added test case) was attached to two edges in the
hypergraph, but with different selectivity on the two edges. So the
cardinality estimate for the join would be different depending on
which of the two edges were used for joining the tables together.

One of the edges (the {t3}-{t1,t2} edge in the test case) used index
statistics for calculating the selectivity, whereas the other
edge (the {t1}-{t3} "cycle" edge) used the fallback selectivity of 0.1
for equality predicates.

The reason why the estimate in the {t1}-{t3} cycle edge did not use
index statistics, was that PromoteCycleJoinPredicates() called
EstimateSelectivity() with an empty companion set argument because
there was no companion set that contained both tables in the cycle
edge. (t1 and t3 are in different companion sets because there is a
straight join between t1 and t2, so both t1 and t2 are alone in their
respective companion sets.) EstimateSelectivity() uses the companion
set to determine if it can use statistics from a compound index.
Lacking this information, it falls back to using a heuristic.

For the other edge, {t3}-{t1,t2}, EstimateSelectivity() is called with
the companion set of the join itself. This companion set does not
include t1 or t2, because of the straight join, but it includes all
the other tables in the query. Based on this companion set,
EstimateSelectivity() is able to see that an index lookup on t3(a, d)
is a possibility for the t1.d = t3.d predicate when combined with the
t3.a = t4.b predicate which exists in the companion set.

Fixed by making AddPredicate() call EstimateSelectivity() with the
companion set of the join to which the predicate belongs, also when
called from PromoteCycleJoinPredicates(), so that
EstimateSelectivity() has the same information and gives the same
result as when it calculates the selectivity for the wider join edge.

Change-Id: Ib3f7681b82af68e861a9250e076fd68e8ba8e752